+++ src/main/kotlin/lotto/utils/ProfitRateCalculator.kt
@@ -3,7 +3,7 @@ package lotto.utils
 import lotto.model.Rank
 
 class ProfitRateCalculator {
-    fun getProfitRate(rankCount: Map<Rank, Int>, price: Int): String {
+    fun calculate(rankCount: Map<Rank, Int>, price: Int): String {
         var profit = 0
         rankCount.forEach { (rank, count) ->
             profit += rank.winningPrice * count
